Move the project sanityChecks to the loadProjectConfig module 🚚
